fix_all

+++ ShoesStore/ShoesStore.Web/Controllers/ProductController.cs
@@ -2,10 +2,13 @@
 using Microsoft.AspNetCore.Mvc;
 using ShoesStore.BLL;
 using ShoesStore.Common.Rsp;
+using ShoesStore.DAL;
+using ShoesStore.DAL.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using System.Web;
 
 namespace ShoesStore.Web.Controllers
 {
@@ -21,20 +24,18 @@ namespace ShoesStore.Web.Controllers
         }
 
         [HttpGet("category/{categoryId}")]
-        public IActionResult GetProductsByCategory(int categoryId)
+        public IActionResult GetProductsByCategoryId(int categoryId)
         {
             var res = new SingleRsp();
-            var products = _productSvc.GetProductsByCategory(categoryId);
-
+            var products = _productSvc.GetProductsByCategoryId(categoryId);
             if (products != null && products.Any())
             {
                 res.Data = products;
             }
             else
             {
-                res.SetError($"No products found in category {categoryId}.");
+                res.SetError("No products found for the specified category.");
             }
-
             return Ok(res);
         }
 
@@ -54,8 +55,7 @@ namespace ShoesStore.Web.Controllers
             return Ok(res);
         }
 
-
-        [HttpGet]
+        [HttpGet("all")]
         public IActionResult GetAllProducts()
         {
             var res = new SingleRsp();
@@ -63,7 +63,29 @@ namespace ShoesStore.Web.Controllers
             return Ok(res);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("sorted-by-price-ascending")]
+        public IActionResult GetProductsSortedByPriceAscending()
+        {
+            var products = _productSvc.GetProductsSortedByPriceAscending();
+            return Ok(products);
+        }
+
+        [HttpGet("sorted-by-price-descending")]
+        public IActionResult GetProductsSortedByPriceDescending()
+        {
+            var products = _productSvc.GetProductsSortedByPriceDescending();
+            return Ok(products);
+        }
+
+        [HttpGet("find/{name}")]
+        public IActionResult GetProductsByName(String name)
+        {
+            var res = new SingleRsp();
+            var products = _productSvc.GetProductsByName(name);
+            return Ok(products);
+        }
+
+        [HttpGet("find/{id:int}")]
         public IActionResult GetProductById(int id)
         {
             var res = new SingleRsp();
@@ -79,23 +101,63 @@ namespace ShoesStore.Web.Controllers
             return Ok(res);
         }
 
-        [HttpPost]
+        [HttpGet("get-newest-products")]
+        public IActionResult GetNewestProducts()
+        {
+            var res = new SingleRsp();
+            var products = _productSvc.GetNewestProducts();
+            return Ok(products);
+        }
+
+        [HttpGet("get-products-from-{lowest}-to-{highest}")]
+        public IActionResult GetNewestProducts(decimal lowest, decimal highest)
+        {
+            var res = new SingleRsp();
+            var products = _productSvc.GetProductsByPriceRange(lowest, highest);
+            return Ok(products);
+        }
+
+        [HttpPost("Create")]
         public IActionResult CreateProduct(string name, string description, decimal price, int quantity, int? categoryId, string imageUrl)
         {
+            if (!AuthRep.IsUserLoggedIn(HttpContext))
+            {
+                return Unauthorized("User is not logged in.");
+            }
+            if (!AuthRep.IsUserAuthorized(HttpContext, 1))
+            {
+                return StatusCode(StatusCodes.Status403Forbidden, "User is not authorized to access this resource.");
+            }
             var res = _productSvc.CreateProduct(name, description, price, quantity, categoryId, imageUrl);
             return Ok(res);
         }
 
-        [HttpPut("{id}")]
+        [HttpPut("Update/{id}")]
         public IActionResult UpdateProduct(int id, string name, string description, decimal price, int quantity, int? categoryId, string imageUrl)
         {
+            if (!AuthRep.IsUserLoggedIn(HttpContext))
+            {
+                return Unauthorized("User is not logged in.");
+            }
+            if (!AuthRep.IsUserAuthorized(HttpContext, 1))
+            {
+                return StatusCode(StatusCodes.Status403Forbidden, "User is not authorized to access this resource.");
+            }
             var res = _productSvc.UpdateProduct(id, name, description, price, quantity, categoryId, imageUrl);
             return Ok(res);
         }
 
-        [HttpDelete("{id}")]
+        [HttpDelete("Delete/{id}")]
         public IActionResult DeleteProduct(int id)
         {
+            if (!AuthRep.IsUserLoggedIn(HttpContext))
+            {
+                return Unauthorized("User is not logged in.");
+            }
+            if (!AuthRep.IsUserAuthorized(HttpContext, 1))
+            {
+                return StatusCode(StatusCodes.Status403Forbidden, "User is not authorized to access this resource.");
+            }
             var res = _productSvc.DeleteProduct(id);
             return Ok(res);
         }
